detection/requirements.txt
# Детекция
numpy               #
scikit-learn        #
scikit-build        #
ultralytics         # YOLOv8
# Обработка фото
opencv-python       # Для работы с изображениями
pillow              # Для обработки изображений
# Сервер
fastapi             # Для создания REST API
uvicorn             # Для запуска сервера
pydantic            #
python-multipart    #
python-dotenv       #
